+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2024-12-02 13:48:27
+
+### Added
+
+- State level subsets.
+- Huggingface token CLI entry.
+
 ## [2.1.0] - 2024-12-02 12:12:59
 
 ### Added
@@ -32,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[2.2.0]: https://github.com/PolicyEngine/policyengine.py/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/PolicyEngine/policyengine.py/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/PolicyEngine/policyengine.py/compare/1.0.1...2.0.0
 [1.0.1]: https://github.com/PolicyEngine/policyengine.py/compare/1.0.0...1.0.1